Create a standalone frontend app
