+++ .gitignore
@@ -33,3 +33,8 @@ pom.xml.tag
 pom.xml.releaseBackup
 pom.xml.versionsBackup
 release.properties
+
+# Javascript
+node
+node_modules
+build
