+++ frontend/package.json
@@ -1,0 +1,45 @@
+{
+  "name": "cloudevent-player",
+  "version": "0.0.0",
+  "description": "CloudEvents Player",
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/ruromero/cloudevent-player.git"
+  },
+  "private": true,
+  "dependencies": {
+    "lock-treatment-tool": "^0.4.1",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "react-scripts": "3.2.0",
+    "typeface-roboto": "0.0.75"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "devDependencies": {
+    "@material-ui/core": "^4.7.0",
+    "@material-ui/icons": "^4.5.1",
+    "react-json-view": "^1.19.1",
+    "uuidv4": "^6.0.0"
+  }
+}
