+++ frontend/setupJest.js
@@ -1,0 +1,3 @@
+//setupJest.js or similar file
+// eslint-disable-next-line no-undef
+global.fetch = require("jest-fetch-mock");
