--- src/main/resources/META-INF/resources/index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>CloudEvents Player</title>
-    <!-- Fonts to support Material Design -->
-    <link rel="stylesheet" href="css/vendor/roboto.css" />
-    <!-- Icons to support Material Design -->
-    <link rel="stylesheet" href="css/vendor/material_icons.css" />
-  </head>
-  <body>
-    <div id="main"></div>
-    <script src="js/vendor/react.production-16.11.0.min.js"></script>
-    <script src="js/vendor/react-dom.production-16.11.0.min.js"></script>
-    <script src="js/vendor/material-ui.production-4.5.1.min.js"></script>
-    <script src="js/vendor/babel-6.26.0.min.js"></script>
-    <script src="js/vendor/uuidv4-3.3.3.js"></script>
-    <script src="js/vendor/react-json-view-1.19.1.min.js"></script>
-    <script type="text/babel" src="js/app.js"></script>
-  </body>
-</html>
